Refactor: move invite-list component to components/mydashboard
