+++ src/pages/index.tsx
@@ -5,8 +5,6 @@ export default function Home(): ReactNode {
   return (
     <div>
       <Link href='/mydashboard'>mydashboard</Link>
-      <br />
-      <Link href='/mydashboard/invite-list'>invite-list</Link>
     </div>
   );
 }
